Moved ToastContext to @Molecules/Toast for better organization
